Workaround for a segfault encountered when resetting build point zone info.
The cause appears to be a system library failure: not resetting static
variables when reloading a plug-in. This is supposed to happen under some
circumstances (references to the library are still held), but this isn't
the case here; nor, so far as I can see, are we using RTLD_NODELETE.

The trigger is lastNumZones == g_humanRepeaterMaxZones.integer when it
should be 0. Because of this, level.buildPointZones isn't reallocated; level
has already been cleared in G_InitGame(), so when it is accessed in
G_CalculateBuildPoints, *boom*. NULL dereference.