Allow blending of two different animations
At the moment, this is more of a client side hack. Ideally, the game should set torsoAnim for animations that require the torso (like NSPA_ATTACK1) and use legsAnim only for WALK/RUN/TURN, howwever, doing this would break md3 animations.
The order of the nonsegmented animations enum was changed to simply some of the logic.