allow particles of child particle systems to be launched with (dynamic) normals
currently, particle systems attached to particles do not have valid normal vectors. such particle systems are not able to launch or displace new particles in a direction based on normal vectors (error messages are printed when particles with normal-based particle trajectories are created).

having a normal vector for a mobile particle system can be useful: one might want to displace physically large particles away from a wall, if the particle system is touching that wall (but do not apply displacement if the system is flying). in this case, a normal vector can be defined by the map surface for the wall. one can also create fancy effects with the use of such normal vectors.

new particle script constructs (in particle blocks):
  {velocityType | accelerationType} {last_normal | opportunistic_normal}

the base vector in case of opportunistic_normal is (0,0,0) if the particle is flying, and is a surface normal vector if the particle is in/on a map surface. the base vector in case of last_normal is always a valid direction, though sometimes an arbitrary one (the up direction). in both cases, the initial value of the last/opportunistic normal vector is propagated from the parent particle.

the last/opportunistic vector is a new particle system variable, and the original normal vector is not touched, so this change retains backward-compatibility.