Have the QVMs built when enabled.
Because cmake is not told that lcc is a compiler, there is insufficient
dependency tracking. Consequently, a hack is needed: the QVMs will be
rebuilt when the corresponding native libs are rebuilt.

(For a custom compiler to be configured, cmake must see it. But since it's
compiled by cmake, we have a slight problem...)