high-precision buildable placement
currently, it's possible to initiate the construction of power/creep-bound buildables by trying to place them at the very edge of the power/creep zone. this was wrongly believed to be caused by floating point imprecision.

when buildables are built, they are placed a bit above the ground, and then "gently" nudged down for whatever reason. however, due to the bouncy physics of the buildables, the buildables bounce up even higher, and stop up there for whatever reason.

from this follows a minimal tutorial on how to easily build unpowered telenodes: build a telenode near the upper edge of the power zone, so that the telenode becomes out of range when it "jumps up" (ie., build in a room above the repeater).

perhaps the nudging was supposed to align the buildable's bottom with the floor, while guaranteeing that buildables don't meld into the ground. however, this displace-and-nudge method should not be required to avoid objects "melding", due to the engine's surface clipping epsilons.

taking the above in mind, i have modified the code to precisely place the buildable at the end of the primary trace, and to keep the buildable stationed there. G_CanBuild() now offers to return the number of the ground entity on which the buildable was placed; the buildable placement functions pass that number to G_Build(). in G_Build(), it is assumed that if a ground entity is specified (ie., if the number is not ENTITYNUM_NONE), then the buildable can stably rest on the surface, and so is stationed there. buildable displacement is also removed, as it is deemed unnecessary.

the most obvious additional benefit of the precise buildable placement is that it's likely to cease all cases of alien buildables falling off the ceiling due to some amount of imprecision. furthermore, the withdrawal of buildable displacement during layout loading can decrease the number of cases of saved buildables failing to load due to start-solid errors.