fix distortion of 2D acceleration
"2D acceleration" is the use of the forwardmove and rightmove controls to gain speed in a direction in scenarios where the upmove control does not give you a 3rd axis of freedom (when not swimming, not climbind a ladder, and not noclipping, ie., walking on a surface or falling in air), but may give you secondary movement abilities (crouching and (wall)jumping).

when accelerating in 2D, in many cases, a non-0 upmove value (perceptibly) lowers the acceleration output of the forwardmove and rightmove controls. for example, players accelerate slower in air (and can't reach higher speeds in some sense) with the jump key held down. this is more significant for marauders when executing multiple walljumps while keeping the jump key down, but is generally significant for strafejumping.

the fix is to differentiate between 2D-based and 3D-based acceleration types (eg., walking vs noclipping).

the patch also removes a couple of obscure modifications of the usercmd_t::upmove value, and, as a byproduct, it fixes the case where you are not able to swim up with the jump key if you ran out of stamina.