Improved radar shows buildings for aliens and humans (client-side)
This radar patch was tested heavily in Balance mod and works very well. It
uses new radar blips which show clients and buildings as differently sized,
crisp-edged circles.  Blips fade out with distance.  The colors have been
chosen to be consistent with team GUIs.

The human radar has not been modified as heavily, it does show the larger
blips for clients so they can be differentiated.