introducing gender-dependent translation
in english, pretty much all adjectives and verbs are independent of the
grammatical gender of the subject. this is not the case in some other
languages, for example spanish and ukrainian. this commit introduces a
way to translate messages dependent on the gender of the sentence's
subject via the use of gettext contexts.
as an added bonus, the conditionals regarding gender needed to handle
reflexive pronouns («himself», «herself», «itself») in CG_Obituary are
no more needed and were safely removed, cutting on the volume of
boilerplate code.