Change the console spam between map changes to be developer only.
Most of the usual information is still displayed when Unvanquished is
first launched, which is useful for troubleshooting. However, it's not
useful to print this every single map change.

With this patch, the only output between map changes is for warnings
and errors (unless "com_developer" is not "0"). Now a player can easily
open their console to see messages from the last game!

Based on Amanieu's Tremulous patch:
http://tinyurl.com/bst3xsu