do not disable the depth tests for implicit shaders with SHADER_2D/LIGHTMAP_2D
when shaders are registered from images directly (ie., RegisterShader("textures/file1.png")), the engine has no choice but to fit the shader with default settings. the chosen settings are geared towards use in overlay images, such as icons (technically, the shader's type is set to SHADER_2D (known as LIGHTMAP_2D in the old renderer)). under these settings, the depth tests are disabled.

this behavior, coupled with the fact that an image is used directly as the shader for the explosion effect of the adv. dragoon's barbs, causes barb explosions to be visible through walls.

disabling of the depth tests for implicit shaders (in case of SHADER_2D/LIGHTMAP_2D) is redundant, as depth testing is disabled for all 2D drawing operations, regardless of shader settings. in this change, the "disable depth test" bit is simply dropped. this fixes the barb issue. other side-effects, if any, are not apparent. the proper solution would be to not use any implicit shaders, ie., to define new shaders and link them to images. however, this change helps.

discussed with: gimhael