do not position buildables too far away from their game-specified origin; display buildables even in air
- a bad design decision was: the cgame displayed buildables at a position depending on 2 traces: a box trace and a capsule trace. the end-position of the capsule trace was used if it didn't differ too much from the end-position of the box trace, and otherwise the end-position of the box trace was used.
- a bug was: the cgame displayed only stationary buildables, not ones in air.

ideally, for gameplay reasons, the graphical model should be displayed at exactly the game-specified origin, to best depict the hit-region of the object. but then, however, the graphical model may appear to float above the ground (this would happen on non-axial surfaces), and for aesthetic reasons, the graphical model could be somewhat displaced towards the ground, but not too much. also, graphical-model-displacement makes no sense when the object is in air.

now, only a capsule trace is done (no box trace). also, buildables are displayed in air.