fix the trapper's anticipation algorithm
the algorithm WRONGly stated the capabilities of lockblob projectiles by
- using a WRONG constant: LOCKBLOB_SPEED was defined as 650, but lockblobs were actually fired at a speed of 500 q/s; and
- not accounting for the (retarded, btw) MISSILE_PRESTEP_TIME adjustments.

define LOCKBLOB_SPEED as 500 and use that constant (instead of the literal 500) in the projectile firing code. also, in the anticipation code, account for the said adjustments.

this makes trappers more proficient against dumb (linearly moving) targets, but more exploitable by smart players in some cases.