Combine binary shader permutations into a single file per shader Also add version information to the files so we can force a recompile of shaders when we change the glsl source code.
Reading a huge number of incredibly small files is not good for cold startup time and it clutters the folder.
Doing it this way also allows the removal of the compile progress bars when loading from the shader cache. We also no longer need to read the shader source files when using the shader cache.