Client side range markers.
The client just draws the range markers as part of the corresponding
buildable. This means that no extra entities have to be communicated
from server to client for the range markers.

The downside is that range markers can be culled with their buildable,
but since buildables in radar range are never culled, this is hopefully
acceptable.