Add a way to send a message (centre print) on or shortly after connection.
This is intended to supply, briefly, information about the current game to
players who've just connected; something which the admin has determined that
players need to know.

This adds the following cvars:
• g_mapStartupMessage – a message for all players
• g_mapStartupMessageDelay – delay (in ms) before which it's shown (5000ms)

One server-to-client command:
• /cpd <delay> "<message>" – output the message after the given delay (in ms)

One cgame command:
• /lcp "<message>" – centre-print the given text locally

g_mapStartupMessage is cleared automatically early in map start-up.
It should be set as needed in maprotation.cfg or a script called from there.