Fixed the pot generation scripts.
I changed the -k parameters format so that they follows xgettext’s doc
and work here. The generate_menu_pot.pl was missing some of the strings.
The parsing code had trouble with some multiline functions and missed
some of the MULTI() choices values. As much of the code was duplicated,
I rewrote it with a rexexp approach, making it smaller and hopefully
more systematic.