Improve rcFilterGaps
This fixes a gap in the generated navmesh for nexus6 near the large transparent tube.
The problem was caused by the gap being small and the ledge being curved at an angle too steep to normally be walked on.

It would be preferable if mappers would clip these small gaps in the floor since future projects like accurate foot placement and bullet physics won't be as forgiving.