Stop legs/torso bone count mismatch from causing disconnection.
The problem only occurs, so far as I can tell, when switching between
players on the human team when spectating when the formerly-spectated
player is represented using an MD3 model and the newly-spectated player is
represented using an MD5 model.

This will cause short-term rendering glitches and console spam; but better
that than disconnection.