Rework and fix map rotation listing.
The current map is highlighted in green and prefixed with ‣.
The next map is shown inline and highlighted in yellow.
Map rotation entry numbers are displayed in white.

Maps not part of the rotation are shown inline but unnumbered.
This means that you could see (colours indicated by comments):

   4 plat23-b9
‣    atcshd                       // green
     nano                         // yellow (g_nextMap)
   5 arachnid2

Display of conditional items has been implemented.