Drop all key events generated while loading; forget what's pressed.
Turns out that they're queued for the first frame and are all processed
then, which defeats clearing the button info. And if they're converted from
SDL input events to game input events, the key-up events get lost (which may
be a problem in the new +command handling).

Easier and cleaner just to drop them all.

‘I was typing when it started loading the next map’ command execution was a
manifestation of this buffering problem and is therefore fixed by this
change.