In the alien spawn menu, always list granger (if enabled).
If advanced granger is locked while the player is not alive then the player
would only be able to spawn as dretch despite adv. granger being listed in
the menu (even if it's opened after adv. granger has been locked). This
workaround ensures that there's always a valid builder option.