Initial work on making VMs work inside the engine process
This is not to be used for production but only for development (and
maybe for servers). The main advantage is that debuggers will see the
VM's memory and threads.
A lot of things are wrong in this commit (NativeInProcessInfo, joining
th VM thread, ...) but at least it is a first working PoC.