Refactor and rework obit messages.
Team indicators are now used. A red disc or a cyan disc is displayed before
each player name, and "TEAMMATE" is no longer present (except for
centre-print).

Environmental kills are grouped a little better.

Assist information now appears on the same line (for long messages).
Some assist information is ignored.

Kills by buildings, when reported in short form, show the assistant (if
there is one) first. This is probably not ideal.

☠ is now used in some short-form messages.