Add back indentation of the templates
This does not break the indentation of the source files.
A preprocess step looks at the template block nesting level of each line
and remove the corresponding indentation, before rendering the template.
It also removes the need for the //% style commands.

One small problem is that it doesn't handle the case where we have a
template block opening and closing on the same line. (there are none for
now)