Add an alert and a work around for a potential bug and add a hint for a future optimization.
The bug, if it is one, appears to be corrupting things and is manifesting itself (on my machine at least), by loading heartbeat.ogg with an incorrect and excessive size. This size causes a memory allocation failure (a bad_alloc exception) which causes the app to crash. Other files may yield the problem I'm not sure.

This commit catches the allocation failure to avoid crashing and reports the problem with the particular ogg file that is known to reproduce the issue, at least on my Windows machine.

The patch calculates the bytes read from the file uses it to check a known good size and report if it doesn't match. This is useful for seeing if the bug is still occuring and finding it.

Once the bug is fixed, some of this hack can be removed.