Fix crash (at least on Windows) when loading large ogg sound files.
Large files like heartbeat.ogg were causing crashes on at least Windows builds. This patch hardens the app to out of memory / bad allocation conditions and teaches it to load as much of a sound file as possible in these circumstances rather than crash and to log a warning that playback may be affected when a sound file can't not be loaded completely.

I initially thought that such large files were caused by a bug but now I think that's not true, we just need to handle them better or perhaps just use smaller files.

Also remove what I think is a redundant std::move statement.

Performance optimizations are potentially possible here but I'll defer those to other patches as the changes may be more invasive and are unrelated to what's required to fix the issue being tackled by this commit.