Refactor hive target selection.
The old algorithm generates a list of all valid targets, sorts that list, and
selects the best target from the sorted list. Since we only need the best
target, we can simplify that. Instead of building a list, the new algorithm only
remembers the best target found yet. This runs faster, requires less memory and
gets rid of a global variable.

Signed-off-by: Viech <viech@unvanquished.net>