Stop using leading _ with new functions and variables to address review comments.
This appears to contradict LibRockets apparent existing convention but the
change complies with the review comments and it does seem to make sense to
move on from that convention on the basis that names using leading
underscores are reserved to the C++ Standard.
It also seems to be less useful in function names than member variables.