Idea: use vector for audio data instead of unique_ptr.
new char[] arrays are not resizeable which is often handy and unique_ptrs are not simply copyable which is also handy even if it moveable.
vector is all of these things and vector with move makes for a simple sink type which AudioData kind of is.
This change also implicitly drops the previous [const] char* interface too as const char* implies won't change but we delete the buffer which although is not changing the data as such, deleting a const char* buffer is perhaps surprising to the user if not technically wrong so non const is arguably more preferable.
I haven't had time to extensively test and tweak and performance check this change but seems to work so far enough for review.
This change improves exception safety.
This change may introduce some alignment issues to gain some simplicity, but the overall benefit may be worth while. Things to consider during review,