Add the autogenerated code to git.
This will allow the game to be compiled right out of checkout.
The cbse toolchain should not be required to build the game,
as it is only required when modifying entities. It is not required
to modify existing entities or to simply build the game for packaging
purposes. Feel free to discuss this with me if you have differing
opinions.