Workaround an optimization bug in Clustering code that could lead to a crash
For some reason, a component of center has a NaN value, and if you multiply
NaN * 0, the result is still NaN. This causes a check to fail and allows
meanObj to remain nullptr, which is then dereferenced leading to a segfault.

This commit explicty forces all the components of center to be 0.0f.

Fixes #851