lua: Add various BG_* getters
Now things like weapons, classes, uprades, buildables can be made
available via the unv global. This can be used in place of a datasources
in the future. Further, in the SGame, there won't be datasources, so
this will be the only (and canonical) way to access information about
the game.

You can test this using something like:
/luarocket "print(unv.weapons['rifle'].name)"
/luarocket "print(unv.buildables['trapper'].name)"
/luarocket "print(unv.upgrades['jetpack'].price)"
/luarocket "print(unv.classes['level3'].price)"