Fix infinite stack growth under heavy load on stdout.
Under heavy I/O on stdout, write() returns -1 and sets
errno to EAGAIN. The return value check causes a
warning to the log which again causes a write to
stdout and on we go. This commit ensures we do not
log to stdout when there is a problem with stdout.
It also uses select to wait until stdout is ready
for writing in case EAGAIN is returned.