Don't read from tess.verts and tess.indexes when backing may be write combined memory
We may use (persistent) glMapBufferRange with GL_MAP_WRITE_BIT but no GL_MAP_READ_BIT for the memory backing of tess.verts and tess.indexes if is available. This usually causes the driver to gives us a pointer to memory marked as write combined or non-cacheable.

Reading from write combined memory is very slow in exchange for fast transfer speeds to the gpu.