Make parameter hierarchy fully functional.
The priorities for picking component paramteter values are, from high to low:
Entity's value, dependent's value, dependency's own default.

At any time you can specify 'null' as a value to clear a default or value that
was set with lower priority, so that the user will be required to give the
value later.