make unvanquished dpkdir a submodule, game loadable from repository
This change makes unvanquished dpkdir a submodule, this base package
was completed with missing assets from previously released pk3 and is
enough to run the game and reach the main menu even while missing
packages, from there it's possible to join a server and autodownload
everything from server.

It means you can now run the game and join a server this way:

	git clone https://github.com/Unvanquished/Unvanquished.git
	cd Unvanquished
	mkdir build
	cd build
	cmake ..
	make
	./daemon -pakpath ../pkg -set vm.sgame.type 3 -set vm.cgame.type 3

At this point you will get some warning about some missing stuff like
soundtrack, but the game will run and you will be able to browse the
server list and join a server if server has autodownload enabled and
serves all required file.

This should be enough to bootstrap a complete game.