Permute yet again the librocket destructors.
It turns out skipping Lua shutdown is not so harmless as that subsystem
retains references to some DOM elements which are then never deleted.
This caused problems when objects in dataSourceMap executed callbacks
into these non-deleted DOM elements.

I commented before that shutting down Lua always caused a crash, but I'm
not getting it now. Maybe changing the order of some other thing helped
or maybe it is going to keep crashing.