Add "checked-value" attribute for cvar checkboxes
This hack is intended to be used for r_dynamicLight where the only
advisable values are 0 or 2. The attribute controls what value is
assigned to the cvar when the box is checked.

Also use atoi as a backup when ParseCvarValue<bool> fails, when
converting a cvar value to a checkbox state. It will be more accurate
for legacy cvars.