cg_view: pretend spectator is alive to avoid death animation to be applied on it
the death animation was applied on spectator because the spectator is not alive,
that's why the noclip spectator was crawling the map with third-person view
while cg_thirdPerson is false.

I don't know why the clip spectator was not affected by the bug and is still
using first-person view while cg_thirdPerson is true. This looks to be unintended.