rename CG_NextWeapon_f to CG_SelectNextInventoryItem_f and other things, fix #1179, ref #544
- rename `CG_NextWeapon_f` to `CG_SelectNextInventoryItem_f` and `CG_PrevWeapon_f` to `CG_SelectPrevInventoryItem_f`, fix #1179.
- make `weapprev`/`weapnext` alias of `itemtoggle blaster` to emulate weapon cycling on mouse roll, ref #544.

See this comment: https://github.com/Unvanquished/Unvanquished/issues/544#issuecomment-619299752

The `weapprev`/`weapnext` thing is done the hacky way to be ready for 0.52 release. A clean implementation is welcome. It is less hacky than storing `itemtoggle blaster` in `default.cfg`, at least the public interface is correct from the start.