hack: use negative STAT_MISC for build timer, positive STAT_MISC will be weapon charge
Since weapon charge require build timer to expire first,
we can wait for build timer to not be smaller than zero
then store the weapon charge in the same field.

That's very ugly and will bite someone hard in the back in the future,
please fix engine instead. The day a bug surfaces because of this hack,
more people would be hurt than the day we would fix the engine.

Not fixing the engine is like feeding an idol who does not mind about us
and does not deserve it.