Fix context menu regression
The context menu in every mode except alien evolution would not update
description properly. This commit fixes that.

this basically corrects
https://github.com/Unvanquished/Unvanquished/pull/1138 and
https://github.com/Unvanquished/Unvanquished/pull/1170. The root problem
was that hiding the element from the formatter did mess with the
indices, and the proposed fix did fix only the menu that broke (the
alien evolve one) while breaking all the others.
This version hides the element properly.

To be merged alongside the corresponding change in unvanquished_src.dpkdir