Zone damage (acid tube) damage rework
https://paste.gnugen.ch/paste/yVlx

I find this makes the acid tubes quite more dangerous. We may want to
reduce the base damage a bit more, then.

The intent of this commit is to avoid the situation where an human sits
at the border of the zone of action, and it looks like the acid tube is
doing something, while it really deals 1 damage per minute.