Fix g_emoticonsAllowedInNames implementation
It looks like it was probably broken during a refactoring to use
Color::Parser.

Also change the method of disrupting the emoticon parsing. Previously it
was to changed "[" to "[["; instead insert a color code (matching the
last used color) after the bracket. The variable names implied that [[
"escapes" the emoji but if there was any cgame code to interpret that as
a single bracket, I think it was lost.

BG_LoadEmoticons still needs to be fixed for the cvar to have any
effect.