Use BG_Emoticon() to get emoticon image paths
Also change the parsing algorithm. Now it runs in a single pass instead
of colors then emojis. This should be more efficient and it will no
longer allow "[^*granger]" as an emoji. Also, any substring comprising a
valid emoticon is recognized, unaffected by previous unmatched '['s.

I removed the restrictions on characters inside the name, instead
relying on the BG_Emoticon database (which implicitly prohibits most of
the characters anyway because of pak filename restrictions).