Unlink player changing team from BSP tree
There was a bug where a bot moved to spectators via an admin command will
never be unlinked. For non-disconnecting clients the unlinking usually
happened during ClientThink of the next frame, but this does not occur
for bot spectators. In any case, it seems better to do this immediately
rather than waiting for the next frame's thinking.