Fix prediction errors on buildables (#425)
The movement code considers an entity to be "on the ground" if it is
within about .2 units above the world/other lower entity. Sometimes when
coming down to a surface, a player may stop, up to 0.2 units above the
surface, without actually touching it. At that point the player may move
about on the flat surface, continuing to "hover" until getting off the
surface.

For unknown reasons, the cgame may predict that the player should fall
down further at that point. This means prediction errors every frame
that the player stays on the surface.

Fix the situation by moving the player down to the surface if the ground
trace fraction is between 0 and 1. This is a similar operation to the
step which may occur if the fraction is 1.