Import tinygettext and build it into cgame
This is based on https://github.com/tinygettext/tinygettext
commit 51d1adc139568ca8bb50a9eecaf62ac0fbfa9803, with a
couple patches that I'm attempting to upstream:
https://github.com/tinygettext/tinygettext/pull/34

I was considering making it a submodule, but tinygettext, bizarrely,
itself has a submodule, which IMO makes the jank too much.

The Daemon submodule is updated here so that there won't be 2
definitions of tinygettext.