Get translation working for some things
Translation now works again for sgame->cgame messages using the print_tr
family of commands, and for some UI strings generated in the cgame, for
example tutorial text. But any fixed strings in RML are not translated.
The translation dictionary now lives in the gamelogic, so it is not
necessary to do syscalls for translation.

src/cgame/translation.cpp is copied from
daemon/src/engine/qcommon/transaltion.cpp with a few changes:
- Remove findlocale part (which still lives in the engine)
- Use gamelogic APIs instead of engine ones
- Restore the functionality of DaemonInputbuf. In the engine, the
  function that does the work was mistakenly deleted in 2015.